app/requirements.txt
prometheus-client
python-dotenv
requests
loguru